fs-tests: perf: fix compilation warning
Fix the following compilation warning:
perf.c: In function ‘perf’:
perf.c:144: warning: format ‘%lld’ expects type ‘long long int’, but argument 2 has type ‘int64_t’

by adding a (long long int) cast.

+++ tests/fs-tests/simple/perf.c
@@ -141,7 +141,8 @@ void perf(void)
 	/* Display timings */
 	printf("File system read and write speed\n");
 	printf("================================\n");
-	printf("Specfied file size: %lld\n", tests_size_parameter);
+	printf("Specfied file size: %lld\n",
+	       (unsigned long long)tests_size_parameter);
 	printf("Actual file size: %zu\n", actual_size);
 	printf("Write time (us): %lld\n", write_time);
 	printf("Unmount time (us): %lld\n", unmount_time);
